requirements.txt
appnope==0.1.3
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.27.6
decorator==5.1.1
Django==4.2.2
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
durationpy==0.5
executing==1.2.0
gunicorn==20.1.0
hjson==3.1.0
idna==3.4
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
placebo==0.9.0
platformdirs==3.8.0
pluggy==1.2.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
djangorestframework_simplejwt==5.3.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
traitlets==5.9.0
troposphere==4.3.2
urllib3==1.26.16
uWSGI==2.0.21
wcwidth==0.2.6
Werkzeug==2.3.6
zappa==0.57.0
python-dotenv==1.0.0
